Separate ContactForm component
